pyproject.toml
[tool.poetry]
name = "puer_analytics"
version = "0.1.0"
description = ""
authors = ["Oleksii Zaichenko <oleksii@shipxdata.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
